refactor: consolidate auth models
