+++ src/ai_karen.egg-info/SOURCES.txt
@@ -128,7 +128,6 @@ src/ai_karen_engine/database/migrations.py
 src/ai_karen_engine/database/models.py
 src/ai_karen_engine/database/schema_validator.py
 src/ai_karen_engine/database/tenant_manager.py
-src/ai_karen_engine/database/models/auth_models.py
 src/ai_karen_engine/doc_store/__init__.py
 src/ai_karen_engine/doc_store/document_store.py
 src/ai_karen_engine/echocore/__init__.py
